requirements.txt
# Core Framework
langchain==0.1.10
langchain-openai==0.0.8
langchain-pinecone==0.0.3
langchain-community==0.0.25

# Vector Store
pinecone-client==3.1.0

# LLM & Embeddings
openai==1.13.3

# UI Framework
streamlit==1.31.1

# Utilities
python-dotenv==1.0.1
requests==2.31.0
tiktoken==0.6.0

# Data Processing
numpy==1.26.4
pandas==2.2.1

